Bug fix
There was a bug associated with the number of allowed tokens. Since how the token count changed with the new ChatGPT model, this system was underestimating the tokens used. A band-aid was applied (decreased from 4000 to 3000 of maximum tokens per request). Hopefully it will be enough.

Fixed a bug with the new search engine where it would return None for certain questions, breaking the code. This is now properly handled.